Сделал доработки
