+++ src/main/java/Controller/DepartmentController.java
@@ -1,5 +1,7 @@
-package com.example.HomeworkLibrary;
+package Controller;
 
+import Service.DepartmentService;
+import Dto.Employee;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
