+++ src/main/java/Exception/EmployeeAlreadyAddedException.java
@@ -1,11 +1,11 @@
-package com.example.HomeworkLibrary;
+package Exception;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
 
 @ResponseStatus(HttpStatus.BAD_REQUEST)
-public class EmployeeNotFoundException extends RuntimeException {
-    public EmployeeNotFoundException(String massage)  {
+public class EmployeeAlreadyAddedException extends RuntimeException{
+    public EmployeeAlreadyAddedException(String massage) {
         super(massage);
     }
 }
