+++ src/main/java/Exception/EmployeeNotFoundException.java
@@ -1,11 +1,11 @@
-package com.example.HomeworkLibrary;
+package Exception;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
 
 @ResponseStatus(HttpStatus.BAD_REQUEST)
-public class EmployeeStorageIsFullException extends RuntimeException {
-    public EmployeeStorageIsFullException(String massage) {
+public class EmployeeNotFoundException extends RuntimeException {
+    public EmployeeNotFoundException(String massage)  {
         super(massage);
     }
 }
