+++ src/main/java/Exception/EmployeeStorageIsFullException.java
@@ -1,11 +1,11 @@
-package com.example.HomeworkLibrary;
+package Exception;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
 
 @ResponseStatus(HttpStatus.BAD_REQUEST)
-public class EmployeeAlreadyAddedException extends RuntimeException{
-    public EmployeeAlreadyAddedException(String massage) {
+public class EmployeeStorageIsFullException extends RuntimeException {
+    public EmployeeStorageIsFullException(String massage) {
         super(massage);
     }
 }
