+++ src/main/java/Exception/IllegalNameException.java
@@ -1,4 +1,4 @@
-package com.example.HomeworkLibrary;
+package Exception;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
