+++ src/main/java/Service/DepartmentService.java
@@ -1,4 +1,6 @@
-package com.example.HomeworkLibrary;
+package Service;
+
+import Dto.Employee;
 
 import java.util.Collection;
 import java.util.List;
