+++ src/main/java/Service/DepartmentServiceImpl.java
@@ -1,5 +1,7 @@
-package com.example.HomeworkLibrary;
+package Service;
 
+import Dto.Employee;
+import Exception.EmployeeNotFoundException;
 import org.springframework.stereotype.Service;
 
 import java.util.Collection;
