+++ src/main/java/Service/EmployeeService.java
@@ -1,4 +1,6 @@
-package com.example.HomeworkLibrary;
+package Service;
+
+import Dto.Employee;
 
 import java.util.Collection;
