+++ src/main/java/Service/HomeworkLibraryApplication.java
@@ -1,4 +1,4 @@
-package com.example.HomeworkLibrary;
+package Service;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
